Fix modules being redefined
So, I didn't understand how modules get defined in angular, and I was
basically re-creating the adn module for each service that was
getting exported.

This fixes that be creating an init.js file that create the module
and all other modules just modify the existing adn module.

I also added more tests for the Auth module.