Adding new components
We are adding some new components to angular-adn including things other
then services. To properly encapsulate everything we had to change
the directory structure.

The new directive is the prototype of an App.net post box. It allows
you to create a post box, with a submit button that will only fire
if everything is with in parameters.